Moved tinyMCE to 6.8.6, sorted public-api file, and removed duplicate function
