+++ package.json
@@ -123,7 +123,7 @@
     "npm-run-all": "^4.1.5",
     "postcss": "^8.4.49",
     "shx": "^0.3.4",
-    "tinymce": "^7.2.1",
+    "tinymce": "^6.8.6",
     "ts-node": "~10.9.2",
     "typescript": "^5.4.5",
     "webpack": "^5.96.1"
