--- packages/angular-sdk-components/src/lib/_components/template/field-group-template/utils.ts
@@ -1,9 +1,0 @@
-export const evaluateAllowRowAction = (allowRowDelete, rowData) => {
-  if (allowRowDelete === undefined || allowRowDelete === true) return true;
-  if (allowRowDelete.startsWith?.('@E ')) {
-    const expression = allowRowDelete.replace('@E ', '');
-    // @ts-ignore
-    return PCore.getExpressionEngine().evaluate(expression, rowData);
-  }
-  return false;
-};
