+++ packages/angular-sdk-components/src/lib/_components/template/utils.ts
@@ -21,3 +21,19 @@ export function filterForFieldValueList(fields: any) {
       value
     }));
 }
+
+/**
+ * This method evaluates whether a row action is allowed based on the provided conditions.
+ * @param {string|boolean|undefined} allowRowDelete - The condition for allowing row deletion.
+ * @param {object} rowData - The data of the row being evaluated.
+ * @returns {boolean} - Returns true if the row action is allowed, false otherwise.
+ */
+export const evaluateAllowRowAction = (allowRowDelete, rowData) => {
+  if (allowRowDelete === undefined || allowRowDelete === true) return true;
+  if (allowRowDelete.startsWith?.('@E ')) {
+    const expression = allowRowDelete.replace('@E ', '');
+    // @ts-ignore - Expected 3 arguments, but got 2
+    return PCore.getExpressionEngine().evaluate(expression, rowData);
+  }
+  return false;
+};
